requirements.txt
django>=5.0
django-ninja>=1.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
import-linter>=2.0
